+++ sqm-render-postgresql/src/main/java/io/sqm/render/pgsql/spi/PostgresBooleans.java
@@ -1,6 +1,0 @@
-package io.sqm.render.pgsql.spi;
-
-import io.sqm.render.ansi.spi.AnsiBooleans;
-
-public class PostgresBooleans extends AnsiBooleans {
-}