+++ sqm-render-postgresql/src/main/java/io/sqm/render/postgresql/ArrayExprRenderer.java
@@ -1,0 +1,34 @@
+package io.sqm.render.postgresql;
+
+import io.sqm.core.ArrayExpr;
+import io.sqm.render.SqlWriter;
+import io.sqm.render.spi.RenderContext;
+import io.sqm.render.spi.Renderer;
+
+public class ArrayExprRenderer implements Renderer<ArrayExpr> {
+    /**
+     * Renders the node into an {@link SqlWriter}.
+     *
+     * @param node a node to render.
+     * @param ctx  a render context.
+     * @param w    a writer.
+     */
+    @Override
+    public void render(ArrayExpr node, RenderContext ctx, SqlWriter w) {
+        w.append("ARRAY").append("[");
+        if (!node.elements().isEmpty()) {
+            w.comma(node.elements());
+        }
+        w.append("]");
+    }
+
+    /**
+     * Gets the target type this handler can handle.
+     *
+     * @return an entity type to be handled by the handler.
+     */
+    @Override
+    public Class<? extends ArrayExpr> targetType() {
+        return ArrayExpr.class;
+    }
+}