+++ sqm-render-postgresql/src/main/java/io/sqm/render/postgresql/spi/PostgresBooleans.java
@@ -1,0 +1,6 @@
+package io.sqm.render.postgresql.spi;
+
+import io.sqm.render.ansi.spi.AnsiBooleans;
+
+public class PostgresBooleans extends AnsiBooleans {
+}