+++ sqm-render-postgresql/src/main/java/io/sqm/render/postgresql/spi/PostgresOperators.java
@@ -1,0 +1,6 @@
+package io.sqm.render.postgresql.spi;
+
+import io.sqm.render.defaults.DefaultOperators;
+
+public class PostgresOperators extends DefaultOperators {
+}